requirements.txt
fastapi[standard]
google-genai
sqlalchemy
psycopg2-binary
